Separated Custom Token Class. A separate file contains the class now for better implementation.
